chore: eliminate actionable build, test, and CI warnings
Migrate UI tests to androidx.compose.ui.test.v2.runComposeUiTest. The v2
suspending block replaces the nested runTest workarounds: setup moves
inside the test block, suspend database calls run directly, and the new
waitUntilSuspending helper polls suspend APIs without blocking, which
wasmJs cannot do.

Returning the v2 TestResult from the test methods makes the wasm browser
tests actually await their bodies, which they silently never did before.
This exposed that compose resources were never served to Karma: add the
karma.config.d resource setup and test_setup.js used by JetBrains in
compose-multiplatform itself, and click confirmation dialog buttons by
test tag instead of localized text.

Also:
- drop a dead elvis branch in OAuthLauncher.ios and a redundant toByte
  conversion in PkceGenerator.wasmJs
- enable native access for desktop test and run JVMs so the Skiko
  System::load warning does not become an error on a future JDK
- silence webpack performance hints (multi-megabyte wasm is inherent)
  and satisfy the memfs tslib peer dependency
- replace the deprecated gradle-home-cache-cleanup setup-gradle input
  with cache-cleanup

Closes #190
Closes #191

+++ composeApp/src/wasmJsMain/kotlin/proj/memorchess/axl/core/auth/PkceGenerator.wasmJs.kt
@@ -14,7 +14,7 @@ internal actual fun randomBytes(size: Int): ByteArray {
   cryptoGetRandomValues(buffer)
   val result = ByteArray(size)
   for (i in 0 until size) {
-    result[i] = buffer[i].toByte()
+    result[i] = buffer[i]
   }
   return result
 }
